+++ deploy/scoper.inc.php
@@ -40,6 +40,17 @@ $polyfillsStubs = array_map(
     ),
 );
 
+$discoverySkipFiles = array_map(
+    static fn (SplFileInfo $fileInfo) => $fileInfo->getPathname(),
+    iterator_to_array(
+        Finder::create()
+            ->files()
+            ->in(dirname(__DIR__))
+            ->name('.discovery-skip'),
+        false,
+    ),
+);
+
 return [
     // The prefix configuration. If a non null value is be used, a random prefix
     // will be generated instead.
@@ -63,6 +74,7 @@ return [
     'exclude-files' => [
         ...$polyfillsBootstraps,
         ...$polyfillsStubs,
+        ...$discoverySkipFiles,
     ],
 
     // When scoping PHP files, there will be scenarios where some of the code being scoped indirectly references the