+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- The `.discovery-skip` files are now included in the release package to ensure the Discovery feature works correctly.
+
 ## [1.0.1] - 2025-12-27
 
 ### Fixed